Converted as helper function and fixed _example.yml comments
