+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/EcoPetsPlugin.kt
@@ -29,7 +29,6 @@ import com.willfp.ecopets.pets.activePet
 import com.willfp.ecopets.pets.activePetLevel
 import com.willfp.ecopets.pets.entity.ModelEnginePetEntity
 import com.willfp.ecopets.pets.entity.PetEntity
-import com.willfp.ecopets.pets.entity.ItemPetEntity
 import com.willfp.ecopets.pets.hasPet
 import com.willfp.libreforge.SimpleProvidedHolder
 import com.willfp.libreforge.conditions.Conditions
@@ -74,10 +73,6 @@ class EcoPetsPlugin : LibreforgePlugin() {
         Filters.register(FilterPet)
         Mutators.register(MutatorPlayerToPet)
         Mutators.register(MutatorPlayerToPetLocation)
-        
-        PetEntity.registerPetEntity("item") { pet, id ->
-            ItemPetEntity(pet, id)
-        }
 
         registerSpecificHolderProvider<Player> {
             it.activePetLevel?.let { p ->
