+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/pets/entity/PetEntity.kt
@@ -7,6 +7,7 @@ import org.bukkit.attribute.Attribute
 import org.bukkit.entity.ArmorStand
 import org.bukkit.entity.Entity
 import org.bukkit.entity.EntityType
+import org.bukkit.entity.ItemDisplay
 import org.bukkit.inventory.EquipmentSlot
 import org.bukkit.inventory.ItemStack
 import org.bukkit.Material
@@ -19,6 +20,10 @@ abstract class PetEntity(
 
     companion object {
         private val registrations = mutableMapOf<String, (Pet, String) -> PetEntity>()
+
+        init {
+            registrations["item"] = ::itemPetEntity
+        }
 
         @JvmStatic
         fun registerPetEntity(id: String, parse: (Pet, String) -> PetEntity) {
@@ -94,5 +99,38 @@ internal fun lookupItem(id: String): ItemStack {
         Items.lookup(id).item ?: ItemStack(Material.BARRIER)
     } catch (e: Exception) {
         ItemStack(Material.BARRIER)
+    }
+}
+
+private fun itemPetEntity(pet: Pet, itemLookup: String): PetEntity = object : PetEntity(pet) {
+    override fun spawn(location: Location): Entity {
+        val item = lookupItem(itemLookup)
+        return if (plugin.configYml.getBool("pet-entity.item-display.enabled")) {
+            spawnAsItemDisplay(location, item)
+        } else {
+            spawnAsArmorStand(location, item)
+        }
+    }
+
+    private fun spawnAsItemDisplay(location: Location, item: ItemStack): ItemDisplay {
+        return location.world!!.spawn(location, ItemDisplay::class.java) {
+            it.setItemStack(item)
+            it.isCustomNameVisible = true
+            it.customName = pet.name
+            it.teleportDuration = plugin.configYml.getInt("pet-entity.item-display.teleport-duration", 3)
+
+            val scale = plugin.configYml.getDouble("pet-entity.scale")
+            if (scale in 0.0625..16.0) {
+                val transform = it.transformation
+                transform.scale.set(scale, scale, scale)
+                it.transformation = transform
+            }
+        }
+    }
+
+    private fun spawnAsArmorStand(location: Location, item: ItemStack): Entity {
+        val stand = emptyArmorStandAt(location, pet, isSkull = true)
+        stand.equipment.helmet = item
+        return stand
     }
 }
