+++ eco-core/core-plugin/src/main/resources/pets/_example.yml
@@ -96,12 +96,14 @@ effects:
 conditions: [ ]
 
 # The texture of the pet entity in game
-# If you're using modelengine, use modelengine:id as the texture
-# If you're using a custom item, use item:id as the texture (e.g. item:diamond_sword)
+# Option 1: A base64 skull texture value
+# Option 2: A vanilla or custom item (prefix with item:)
+# entity-texture: item:diamond_sword
+# Option 3: A ModelEngine model (prefix with modelengine:)
+# entity-texture: modelengine:my_model
 entity-texture: "eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvOTA5NWZjYzFlM2Q3Y2JkMzUwZjE5YjM4OTQ5OGFiOGJiOTZjNjVhZDE4NWQzNDU5MjA2N2E3ZDAzM2FjNDhkZSJ9fX0="
 
 # The icon in GUIs
-# You can use eco items here too! E.g. icon: diamond_sword
 icon: player_head texture:eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvOTA5NWZjYzFlM2Q3Y2JkMzUwZjE5YjM4OTQ5OGFiOGJiOTZjNjVhZDE4NWQzNDU5MjA2N2E3ZDAzM2FjNDhkZSJ9fX0=
 
 # The spawn egg
